99kdumpbase: Fix attempting to setup kdump network if no network interface is used
kdump_install_net() also attempts to setup kdump network if no network
interface is used.

This is because _get_kdump_netifs() always outputs not empty string.
When both unique_netifs and ovs_unique_netifs are empty,
_get_kdump_netifs() incorrectly outputs " " (one space).
In that case, it is expected to output empty string.

This may cause that unnecessary network-related files are installed to
initramfs.

This patch fixes this issue by correcting _get_kdump_netifs() output.

Fixes: 3f4a9eb4dbf9 ("Bring up the network before Open vSwitch bridge")
Signed-off-by: Kazuma Kondo <kazuma-kondo@nec.com>